+++ test/models/ui/create_account_model_test.dart
@@ -1,5 +1,5 @@
 import "package:flutter_test/flutter_test.dart";
-import "package:proxima/models/ui/create_account_validation.dart";
+import "package:proxima/models/ui/validation/create_account_validation.dart";
 
 void main() {
   group("Create Account Model testing", () {
