+++ test/models/ui/new_post_state_test.dart
@@ -1,5 +1,5 @@
 import "package:flutter_test/flutter_test.dart";
-import "package:proxima/models/ui/new_post_validation.dart";
+import "package:proxima/models/ui/validation/new_post_validation.dart";
 
 void main() {
   group("Create Account Model testing", () {
